+++ platform/linuxbsd/display_server_x11.cpp
@@ -2212,8 +2212,8 @@ void DisplayServerX11::_handle_key_event(WindowID p_window, XKeyEvent *p_event, 
 	// XLookupString returns keysyms usable as nice keycodes.
 	char str[256 + 1];
 	XKeyEvent xkeyevent_no_mod = *xkeyevent;
-	xkeyevent_no_mod.state &= ~ShiftMask;
-	xkeyevent_no_mod.state &= ~ControlMask;
+	xkeyevent_no_mod.state &= ~(uint32_t(ShiftMask));
+	xkeyevent_no_mod.state &= ~(uint32_t(ControlMask));
 	XLookupString(xkeyevent, str, 256, &keysym_unicode, nullptr);
 	XLookupString(&xkeyevent_no_mod, nullptr, 0, &keysym_keycode, nullptr);
